pyproject.toml
[tool.poetry]
name = "skill-manager"
version = "0.1.0"
description = ""
authors = ["Tim <jlemyp96@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
django = "^2.2"
djangorestframework_simplejwt = "^4.3"
django-reversion = "^3.0"
django-filter = "^2.2"
django-extensions = "^2.2"
pyyaml = "^5.1"
uritemplate = "^3.0"
django-loginas = "^0.3.6"
django-cors-headers = "^3.1"
django-constance = {version = "^2.4",extras = ["database"]}
djangorestframework-serializer-extensions = "^1.0"
django-autocompletefilter = "^0.0.5"
django-split-settings = "^1.0"
djangorestframework = "^3.10"
# Pillow = "^6.2"


django-environ = "^0.4.5"
[tool.poetry.dev-dependencies]
Werkzeug = "^0.15.5"
ipython = "^7.7"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
